Ep 136.

+++ code/handmade_sim_region.h
@@ -85,7 +85,7 @@ struct sim_entity
 
     sim_entity_collision_volume_group *Collision;
 
-    uint32 FacingDirection;
+    real32 FacingDirection;
     real32 tBob;
 
     int32 dAbsTileZ;
